* New UI and condition building system for filter builder. The idea is to make the interface as intuitive as possible. Whether this is achieved here is of course open to criticism. * Renamed **FilterBuilderView[Model]** to **FilterManagerView[Model]**, I personnaly always got confused about its content. * Added new **FilterBuilderView[Model]** which handles display and editing calls for building **FilterScheme**. * Moved **EditFilterView[Model]** filter-building part to said new **FilterBuilderView[Model]** couple (well, triple .xaml[.cs]) * Added filter tree building logic to **FilterSchemeExtensions.cs** * Added **StackTreeView.xaml** which defines the new styling items for TreeView filter rendering * Added **FilterBuilderGrayBrush#** color brushes to **Generic.xaml** to style the TreeView colors
