+++ src/Orc.FilterBuilder.Example/Views/RibbonView.xaml
@@ -14,7 +14,7 @@
                     <fluent:SplitButton Header="{Binding SelectedFilterScheme.Title}" Command="{Binding NewSchemeCommand}"
                                         LargeIcon="{orchestra:FontImage {x:Static ribbon:FontAwesome.Plus}}" 
                                         ToolTip="{Binding SelectedFilterScheme}" ClosePopupOnMouseDown="True">
-                        <views:FilterBuilderControl RawCollection="{Binding RawItems}" CloseViewModelOnUnloaded="False" />
+                        <views:FilterManagerControl RawCollection="{Binding RawItems}" CloseViewModelOnUnloaded="False" />
                     </fluent:SplitButton>
 
                 </fluent:RibbonGroupBox>
