+++ src/Orc.FilterBuilder/Orc.FilterBuilder.Shared/FilterBuilderMode.cs
@@ -10,17 +10,17 @@ namespace Orc.FilterBuilder
     using Views;
 
     /// <summary>
-    /// Enumeration of <see cref="FilterBuilderControl"/> work modes
+    /// Enumeration of <see cref="FilterManagerControl"/> work modes
     /// </summary>
     public enum FilterBuilderMode
     {
         /// <summary>
-        /// <see cref="FilterBuilderControl"/> creates filtered collection
+        /// <see cref="FilterManagerControl"/> creates filtered collection
         /// </summary>
         Collection,
 
         /// <summary>
-        /// <see cref="FilterBuilderControl"/> creates filtering function
+        /// <see cref="FilterManagerControl"/> creates filtering function
         /// </summary>
         FilteringFunction
     }
