+++ src/Orc.FilterBuilder/Orc.FilterBuilder.Shared/Views/DataTemplates/ConditionGroupView.xaml
@@ -1,0 +1,45 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:catel="http://catel.codeplex.com"
+                    xmlns:filterBuilder="clr-namespace:Orc.FilterBuilder"
+                    xmlns:markup="clr-namespace:Orc.FilterBuilder.Markup">
+
+  <HierarchicalDataTemplate DataType="{x:Type filterBuilder:ConditionGroup}"
+                            ItemsSource="{Binding Items}">
+
+    <HierarchicalDataTemplate.Resources>
+      <catel:ObjectToDisplayNameConverter x:Key="ObjectToDisplayNameConverter" />
+    </HierarchicalDataTemplate.Resources>
+
+    <Border BorderBrush="{DynamicResource FilterBuilderGrayBrush9}"
+            BorderThickness="1 1 1 1">
+
+      <Grid Background="{DynamicResource FilterBuilderGrayBrush10}"
+            Width="90">
+
+        <Grid Height="60"
+              Margin="0"
+              VerticalAlignment="Top">
+
+          <ComboBox ItemsSource="{markup:EnumBinding {x:Type filterBuilder:ConditionGroupType}}"
+                    SelectedItem="{Binding Type}"
+                    Style="{DynamicResource FilterBuilderComboBoxStyle}"
+                    Width="62"
+                    HorizontalAlignment="Center"
+                    VerticalAlignment="Center">
+            <ComboBox.ItemTemplate>
+              <DataTemplate>
+                <TextBlock Text="{Binding Converter={StaticResource ObjectToDisplayNameConverter}}" />
+              </DataTemplate>
+            </ComboBox.ItemTemplate>
+          </ComboBox>
+
+        </Grid>
+
+      </Grid>
+
+    </Border>
+
+  </HierarchicalDataTemplate>
+
+</ResourceDictionary>
