+++ src/Orc.FilterBuilder/Orc.FilterBuilder.Shared/Views/FilterBuilderView.xaml
@@ -1,0 +1,61 @@
+﻿<catel:UserControl x:Class="Orc.FilterBuilder.Views.FilterBuilderView"
+                   xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                   xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                   xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                   xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                   xmlns:catel="http://catel.codeplex.com">
+
+  <ScrollViewer HorizontalScrollBarVisibility="Auto"
+                VerticalScrollBarVisibility="Auto">
+
+    <StackPanel HorizontalAlignment="Left"
+                VerticalAlignment="Top"
+                Margin="20 20 20 20">
+
+      <StackPanel.Resources>
+        <ResourceDictionary>
+          <ResourceDictionary.MergedDictionaries>
+            <ResourceDictionary
+              Source="pack://application:,,,/Orc.FilterBuilder;component/Views/DataTemplates/ConditionGroupView.xaml" />
+            <ResourceDictionary
+              Source="pack://application:,,,/Orc.FilterBuilder;component/Views/DataTemplates/PropertyExpressionView.xaml" />
+            <ResourceDictionary
+              Source="pack://application:,,,/Orc.FilterBuilder;component/Themes/StackedTreeView.xaml" />
+          </ResourceDictionary.MergedDictionaries>
+        </ResourceDictionary>
+      </StackPanel.Resources>
+
+      <TreeView ItemContainerStyle="{DynamicResource StackedTreeViewItemStyle}"
+                HorizontalContentAlignment="Stretch"
+                VerticalContentAlignment="Stretch"
+                BorderBrush="{DynamicResource FilterBuilderGrayBrush7}"
+                BorderThickness="1"
+                ItemsSource="{Binding ConditionItems}">
+
+        <TreeView.Template>
+          <ControlTemplate TargetType="TreeView">
+            <Grid>
+              <Border BorderBrush="{TemplateBinding BorderBrush}"
+                      BorderThickness="{TemplateBinding BorderThickness}">
+                <ItemsPresenter Margin="0" />
+              </Border>
+            </Grid>
+          </ControlTemplate>
+        </TreeView.Template>
+
+        <!-- Put the root item(s) in a centered Grid so that 
+         they will be centered and retain their width. -->
+        <TreeView.ItemsPanel>
+          <ItemsPanelTemplate>
+            <Grid HorizontalAlignment="Center"
+                  IsItemsHost="True" />
+          </ItemsPanelTemplate>
+        </TreeView.ItemsPanel>
+
+      </TreeView>
+    </StackPanel>
+
+  </ScrollViewer>
+
+</catel:UserControl>
+
