+++ src/Orc.FilterBuilder/Orc.FilterBuilder.Shared/Views/FilterBuilderView.xaml.cs
@@ -1,0 +1,32 @@
+﻿// --------------------------------------------------------------------------------------------------------------------
+// <copyright file="PropertyExpression.cs" company="WildGums">
+//   Copyright (c) 2008 - 2016 WildGums. All rights reserved.
+// </copyright>
+// --------------------------------------------------------------------------------------------------------------------
+
+
+namespace Orc.FilterBuilder.Views
+{
+    using Catel.Windows.Controls;
+
+    /// <summary>
+  ///   Interaction logic for FilterBuilderView.xaml
+  /// </summary>
+  public partial class FilterBuilderView : UserControl
+  {
+    #region Constructors
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="FilterBuilderView"/> class.
+    /// </summary>
+    /// <remarks>
+    /// This method is required for design time support.
+    /// </remarks>
+    public FilterBuilderView()
+    {
+      InitializeComponent();
+    }
+
+    #endregion
+  }
+}
