+++ src/Orc.FilterBuilder/Orc.FilterBuilder.Shared/Views/FilterManagerControl.xaml
@@ -1,4 +1,4 @@
-﻿<catel:UserControl x:Class="Orc.FilterBuilder.Views.FilterBuilderControl"
+﻿<catel:UserControl x:Class="Orc.FilterBuilder.Views.FilterManagerControl"
                    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                    xmlns:catel="http://catel.codeplex.com"
